requirements.txt
# Requirements pour l'application Streamlit
# Composants principaux
streamlit
numpy
pandas
scipy
# Pour le traitement des images
opencv-python-headless
Pillow
# Pour la visualisation
matplotlib
seaborn
# Pour le machine learning
tensorflow
scikit-learn
# Pour le suivi d'expériences
mlflow
# Autres outils utiles
tqdm
ipywidgets
jupyterlab
ipykernel
# Pour le déploiement
Flask
